Move BaseProductList UI to internal segment
