+++ src/entities/category/ui/CategoryCard/CategoryCard.module.css
@@ -20,6 +20,7 @@
   width: 100%;
   height: 100%;
   min-height: 300px;
+  background-color: #7775;
   background-size: cover;
   background-position: center;
   overflow: hidden;
