+++ src/widgets/BaseProductList/ui/BaseProductList.module.css
