+++ src/widgets/BaseProductList/ui/BaseProductList.stories.tsx
